Extract auth details of the Header component to its own sub-component
